[Edit] 스테이지 종료 후 보상 창 수정 및 스테이지 정보 프리펩 추가
스테이지 종료 후 보상UI가 나올 수 있도록 수정 및 스테이지 정보를 받아 계속 나타내는 Stagenfo 프리팹 추가
