+++ Assets/Scripts/UI/StageInfomation.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: be0397f1d2a7fd24eb0124cd43084824
